hunting: implement huntMultiple functionality
This patch implements proper multiple hunting squad functionality. It does
some monkeypatching to extend a proper huntMultiple function and calls it
to perform the hunting. This may not be an acceptable approach. However,
the benefits are such that we get ability to limit hunts from between 1
and all, without having to repeat-click the button (which generates too
many log messages).

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>